retype_dataframe implemented

+++ data_management/src/raw_data/schemas/raw_data_data.py
@@ -4,23 +4,23 @@ import polars as pl
 class AdjustedPricesData:
     unix_date_time: pl.Int64
     symbol: pl.Utf8
-    price: pl.Float32
+    price: pl.Float64
 
 
 class FxPricesData:
     unix_date_time: pl.Int64
     symbol: pl.Utf8
-    price: pl.Float32
+    price: pl.Float64
 
 
 class MultiplePricesData:
     unix_date_time: pl.Int64
     symbol: pl.Utf8
-    carry: pl.Float32
+    carry: pl.Float64
     carry_contract: pl.Int64
-    price: pl.Float32
+    price: pl.Float64
     price_contract: pl.Int64
-    forward: pl.Float32
+    forward: pl.Float64
     forward_contract: pl.Int64
